Dedup one balance data and reduce network calls
Fixing tests on the way too
