+++ api/tests/test_assets.py
@@ -12,21 +12,6 @@ app_client = TestClient(app)
 
 @pytest.fixture()
 def patch_database(monkeypatch):
-    data = [
-        {
-            "_id": "65d39b44a7ff2a5e843b1e0b",
-            "time": "2024-02-19 18:17:40.479",
-            "data": [
-                {
-                    "symbol": "PIXELUSDT",
-                    "priceChangePercent": "15.000",
-                    "volume": "1235816253.30000000",
-                    "price": "0.64880000",
-                }
-            ],
-        }
-    ]
-
     itemized_balance = [
         {"asset": "BTC", "free": 6.51e-06, "locked": 0.0},
         {"asset": "BNB", "free": 9.341e-05, "locked": 0.0},
@@ -38,7 +23,6 @@ def patch_database(monkeypatch):
         self._db = db
 
     monkeypatch.setattr(Assets, "_db", new_init)
-    monkeypatch.setattr(Assets, "get_market_domination", lambda a, size, *arg: data)
     monkeypatch.setattr(Assets, "get_fiat_coin", lambda self: "USDC")
     monkeypatch.setattr(
         Assets, "get_raw_balance", lambda self, asset=None: itemized_balance
